Not append JSON ext to base URL when it's not needed (#14)
* Not include JSON ext to base URL when not needed

Previous versions always included .json extension
to the base url in the Firebase class when building a full URL.

It is indeed needed when working with the Firebase's Database service,
but on the other hand brings issues when working with
the Authentication service API
(https://firebase.google.com/docs/reference/rest/auth/). The urls there
don't need any extensions.

The commit provides the user an opportunity to specify in
the constructor of Firebase class not to add that extension to the URL.

Demo class is updated accordingly; an example of signing up a user
is provided.

* README Modified: example to exclude .json ext

An example is added to README on how to use the overloaded constructor
in order to prevent the .json extension to be added to the base url.